WIP: refactoring and fix bugs
